requirement.txt
astroid==2.0.4
autopep8==1.4
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.4
mccabe==0.6.1
pycodestyle==2.4.0
pylint==2.1.1
python-docx==0.8.7
six==1.11.0
wrapt==1.10.11
yapf==0.23.0
